chore: Move test helper to storefront package
