+++ tests/integration/Storefront/Theme/ThemeInheritanceBuilderTest.php
@@ -5,8 +5,8 @@ namespace Shopware\Tests\Integration\Storefront\Theme;
 use PHPUnit\Framework\TestCase;
 use Shopware\Core\Framework\Bundle;
 use Shopware\Core\Framework\Test\TestCaseBase\IntegrationTestBehaviour;
-use Shopware\Core\Test\Stub\Storefront\ThemeRuntimeConfigTestService;
 use Shopware\Storefront\Storefront;
+use Shopware\Storefront\Test\Theme\ThemeRuntimeConfigTestService;
 use Shopware\Storefront\Theme\StorefrontPluginConfiguration\AbstractStorefrontPluginConfigurationFactory;
 use Shopware\Storefront\Theme\StorefrontPluginConfiguration\StorefrontPluginConfigurationCollection;
 use Shopware\Storefront\Theme\StorefrontPluginConfiguration\StorefrontPluginConfigurationFactory;
